Changed spawner script, added incidents.
