+++ Assets/Prefabs/Incidents/Car Crash Spawner.prefab
@@ -9,9 +9,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2244365628043059125}
-  - component: {fileID: 2056113704761620133}
   - component: {fileID: -1972478628280216212}
   - component: {fileID: -5988451830798414340}
+  - component: {fileID: -5798225353224187968}
   m_Layer: 0
   m_Name: Car Crash Spawner
   m_TagString: Untagged
@@ -34,22 +34,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &2056113704761620133
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1633113725953468390}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0acc8516f9e148647a344d4184d5b481, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  incidentInstigators:
-  - {fileID: 3523345102258385620, guid: 83e329bacac4dc5429bcf4941721794e, type: 3}
-  incidentReplacements: []
-  qnaList: []
 --- !u!23 &-1972478628280216212
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -100,3 +84,18 @@ MeshFilter:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1633113725953468390}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
+--- !u!114 &-5798225353224187968
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1633113725953468390}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9893f21039f846a48af6d3de8bed6025, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  incidentInstigators:
+  - {fileID: 3523345102258385620, guid: 83e329bacac4dc5429bcf4941721794e, type: 3}
+  incidentReplacements: []
