+++ Assets/Prefabs/Incidents/Car Crash.prefab
@@ -9,7 +9,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1896618561391410756}
-  - component: {fileID: -8294232347797164050}
+  - component: {fileID: 850570162835856650}
+  - component: {fileID: 9210169060992658627}
+  - component: {fileID: 1522879966222949289}
   m_Layer: 0
   m_Name: Car Crash
   m_TagString: Untagged
@@ -34,27 +36,84 @@ Transform:
   - {fileID: 8358245528025157496}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!65 &-8294232347797164050
-BoxCollider:
+--- !u!1839735485 &850570162835856650
+Tilemap:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3523345102258385620}
-  m_Material: {fileID: 0}
+  m_Enabled: 1
+  m_Tiles: {}
+  m_AnimatedTiles: {}
+  m_TileAssetArray: []
+  m_TileSpriteArray: []
+  m_TileMatrixArray: []
+  m_TileColorArray: []
+  m_TileObjectToInstantiateArray: []
+  m_AnimationFrameRate: 1
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_Origin: {x: 0, y: 0, z: 0}
+  m_Size: {x: 0, y: 0, z: 1}
+  m_TileAnchor: {x: 0.5, y: 0.5, z: 0}
+  m_TileOrientation: 0
+  m_TileOrientationMatrix:
+    e00: 1
+    e01: 0
+    e02: 0
+    e03: 0
+    e10: 0
+    e11: 1
+    e12: 0
+    e13: 0
+    e20: 0
+    e21: 0
+    e22: 1
+    e23: 0
+    e30: 0
+    e31: 0
+    e32: 0
+    e33: 1
+--- !u!54 &9210169060992658627
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3523345102258385620}
+  serializedVersion: 4
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0
+  m_CenterOfMass: {x: 0, y: 0, z: 0}
+  m_InertiaTensor: {x: 1, y: 1, z: 1}
+  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
   m_IncludeLayers:
     serializedVersion: 2
     m_Bits: 0
   m_ExcludeLayers:
     serializedVersion: 2
     m_Bits: 0
-  m_LayerOverridePriority: 0
-  m_IsTrigger: 0
-  m_ProvidesContacts: 0
+  m_ImplicitCom: 0
+  m_ImplicitTensor: 1
+  m_UseGravity: 0
+  m_IsKinematic: 1
+  m_Interpolate: 0
+  m_Constraints: 0
+  m_CollisionDetection: 0
+--- !u!114 &1522879966222949289
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3523345102258385620}
   m_Enabled: 1
-  serializedVersion: 3
-  m_Size: {x: 1, y: 1, z: 1}
-  m_Center: {x: 0, y: 0, z: 0}
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0acc8516f9e148647a344d4184d5b481, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  qnaList: []
 --- !u!1001 &4790072816597669085
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -110,8 +169,38 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 8327660804240166455, guid: 665c762ebb58ff74f9d01ece1d8e9f64, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 104206802200838045}
   m_SourcePrefab: {fileID: 100100000, guid: 665c762ebb58ff74f9d01ece1d8e9f64, type: 3}
+--- !u!1 &3596136088502196970 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 8327660804240166455, guid: 665c762ebb58ff74f9d01ece1d8e9f64, type: 3}
+  m_PrefabInstance: {fileID: 4790072816597669085}
+  m_PrefabAsset: {fileID: 0}
+--- !u!64 &104206802200838045
+MeshCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3596136088502196970}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 5
+  m_Convex: 0
+  m_CookingOptions: 30
+  m_Mesh: {fileID: 2983147389569101140, guid: 5e5f2f8dfeffd9d4baf7ec4d1d51b212, type: 3}
 --- !u!4 &6776855776091752544 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2050821961311837373, guid: 665c762ebb58ff74f9d01ece1d8e9f64, type: 3}
@@ -176,8 +265,38 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 3991679727971031532, guid: fc26a2b872b51494a9251c604993a2a4, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 334482676225290567}
   m_SourcePrefab: {fileID: 100100000, guid: fc26a2b872b51494a9251c604993a2a4, type: 3}
+--- !u!1 &5144322234019610247 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 3991679727971031532, guid: fc26a2b872b51494a9251c604993a2a4, type: 3}
+  m_PrefabInstance: {fileID: 8070738882296457067}
+  m_PrefabAsset: {fileID: 0}
+--- !u!64 &334482676225290567
+MeshCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5144322234019610247}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 5
+  m_Convex: 0
+  m_CookingOptions: 30
+  m_Mesh: {fileID: -5265205315050602288, guid: 5e5f2f8dfeffd9d4baf7ec4d1d51b212, type: 3}
 --- !u!4 &8358245528025157496 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 288074548855269395, guid: fc26a2b872b51494a9251c604993a2a4, type: 3}
