+++ Assets/Prefabs/Incidents/Lamp Crash Spawner.prefab
@@ -9,9 +9,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2999821288366138042}
-  - component: {fileID: 5874561528578860430}
   - component: {fileID: -7710455693815638958}
   - component: {fileID: 6937967763131454858}
+  - component: {fileID: -3689260317905005366}
   m_Layer: 0
   m_Name: Lamp Crash Spawner
   m_TagString: Untagged
@@ -34,23 +34,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &5874561528578860430
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8577812532067678297}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7f086d94b994422aa534affc6f87054b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  incidentInstigators:
-  - {fileID: 3523345102258385620, guid: dab5c02cdfe7d7a4298e4994567b2426, type: 3}
-  - {fileID: 3523345102258385620, guid: cfaa0e1132f31be45b6350b54b89db6c, type: 3}
-  incidentReplacements:
-  - {fileID: 5182456745956613823, guid: e5d8691e8c25d4d43bd3b2ab5babba14, type: 3}
 --- !u!33 &-7710455693815638958
 MeshFilter:
   m_ObjectHideFlags: 0
@@ -101,3 +84,20 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
+--- !u!114 &-3689260317905005366
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8577812532067678297}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9893f21039f846a48af6d3de8bed6025, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  incidentInstigators:
+  - {fileID: 3523345102258385620, guid: cfaa0e1132f31be45b6350b54b89db6c, type: 3}
+  - {fileID: 3523345102258385620, guid: dab5c02cdfe7d7a4298e4994567b2426, type: 3}
+  incidentReplacements:
+  - {fileID: 5182456745956613823, guid: e5d8691e8c25d4d43bd3b2ab5babba14, type: 3}
