+++ Assets/Scripts/Incidents/IncidentBase.cs
@@ -6,33 +6,18 @@ namespace Incidents
 {
     public abstract class IncidentBase : MonoBehaviour
     {
-        [SerializeField] private GameObject[] incidentInstigators;
-        [SerializeField] private GameObject[] incidentReplacements;
         [SerializeField] public List<QuestionAndAnswers> qnaList;
         int randomIntInRange;
-        bool isAnswered = false;
+        bool isAnswered = false;      
 
-        private void Awake()
-        {
-            GetComponent<MeshRenderer>().enabled = false;
-        }
-        
         private void Start()
         {
-            randomIntInRange = Random.Range(0, qnaList.Count);
+            if (qnaList != null)
+            {
+                randomIntInRange = Random.Range(0, qnaList.Count);
+            }
         }
 
-        public void SpawnIncident()
-        {
-            Instantiate(incidentInstigators[Random.Range(0, incidentInstigators.Length)], transform.position, transform.rotation);
-        }
-
-        public void SpawnReplacement()
-        {
-            if (incidentReplacements != null && incidentReplacements.Length > 0)
-                Instantiate(incidentReplacements[Random.Range(0, incidentReplacements.Length)], transform.position, transform.rotation);
-        }
-        
         public void addQuestionAndAnswer(List<QuestionAndAnswers> qnaList )
         {
             this.qnaList = qnaList;
