+++ Assets/Scripts/SpawnManager.cs
@@ -6,13 +6,13 @@ public class SpawnManager : MonoBehaviour
 {
     void Start()
     {
-        IncidentBase[] allIncidents = FindObjectsByType<IncidentBase>(FindObjectsSortMode.None);
+        IncidentSpawner[] allIncidents = FindObjectsByType<IncidentSpawner>(FindObjectsSortMode.None);
         var groupedIncidents = allIncidents.GroupBy(incident => incident.GetType());
         foreach (var group in groupedIncidents)
         {
-            IncidentBase rndIncident = group.ElementAt(Random.Range(0, group.Count()));
+            IncidentSpawner rndIncident = group.ElementAt(Random.Range(0, group.Count()));
             rndIncident.SpawnIncident();
-            foreach (IncidentBase incident in group.Where(i => i != rndIncident))
+            foreach (IncidentSpawner incident in group.Where(i => i != rndIncident))
             {
                 incident.SpawnReplacement();
             }
